requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
watchfiles>=0.21.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.3

# Computer Vision and Machine Learning
# Use headless for server environments (no UI required)
opencv-python>=4.9.0
numpy>=1.24.0
onnxruntime>=1.15.0
mediapipe>=0.10.0

# AI and External APIs
openai>=1.0.0
httpx>=0.25.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0
motor>=3.3.1
pymupdf>=1.24.0
